+++ modules/types/src/client.ts
@@ -10,7 +10,7 @@ import {
 import { ConnextEvent } from "./basic";
 import { AppInstanceJson, CFCoreTypes } from "./cf";
 import { CFCoreChannel, ChannelAppSequences, ChannelState, PaymentProfile } from "./channel";
-import { ChannelProvider, ChannelRouterConfig } from "./channelProvider";
+import { ChannelProvider, ChannelProviderConfig } from "./channelProvider";
 import {
   ConditionalTransferParameters,
   ConditionalTransferResponse,
@@ -84,7 +84,7 @@ export interface IConnextClient {
   resolveCondition(params: ResolveConditionParameters): Promise<ResolveConditionResponse>;
   conditionalTransfer(params: ConditionalTransferParameters): Promise<ConditionalTransferResponse>;
   restoreState(): Promise<void>;
-  channelProviderConfig(): Promise<ChannelRouterConfig>;
+  channelProviderConfig(): Promise<ChannelProviderConfig>;
   requestDepositRights(
     params: RequestDepositRightsParameters,
   ): Promise<CFCoreTypes.RequestDepositRightsResult>;
